Daily and ui (#18)
* feat(daily calculations):

arima.py:
* moved Repo inside class so the db connection is propery closed
when the Arima class gets deleted
* added an argument 'metric_ids' to get_data method signature in order
to decrease db transaction time

context.py:
* added a property to prediction to make generating the drop down list
more fluid

repository.py:
* alphabetized method order
* added new method to all runs as list rather than dataframe
* added 'metric_ids' as optional argument to get_measurements to reduce
the amount of unnecessary rows retrieved
* added a method 'get_predictions' to pull precomputed predictions from
the database to make ui more responsive
* changed method signature for put_measurements and added another method
so we can put measurements from either a csv file or a list

ui.py: new, a start to the ui

* feat(refactor): updates to daily.py

* feat(refactor): updates to daily.py

* feat(ui): distinguish between past and predicted values

* feat(ui): distinguish between past and predicted values

* feat(ui): distinguish between past and predicted values

* feat(ui): widget functionality

all widgets are working as intended. still need to comment and write a
few unit tests. i may try to write some selenium tests simply because
i never have and this is a small UI. would be a good place to try it out

* feat(daily calculations):

arima.py:
* moved Repo inside class so the db connection is propery closed
when the Arima class gets deleted
* added an argument 'metric_ids' to get_data method signature in order
to decrease db transaction time

context.py:
* added a property to prediction to make generating the drop down list
more fluid

repository.py:
* alphabetized method order
* added new method to all runs as list rather than dataframe
* added 'metric_ids' as optional argument to get_measurements to reduce
the amount of unnecessary rows retrieved
* added a method 'get_predictions' to pull precomputed predictions from
the database to make ui more responsive
* changed method signature for put_measurements and added another method
so we can put measurements from either a csv file or a list

ui.py: new, a start to the ui

* feat(refactor): updates to daily.py

* feat(refactor): updates to daily.py

* feat(ui): distinguish between past and predicted values

* feat(ui): distinguish between past and predicted values

* feat(ui): distinguish between past and predicted values

* feat(ui): widget functionality

all widgets are working as intended. still need to comment and write a
few unit tests. i may try to write some selenium tests simply because
i never have and this is a small UI. would be a good place to try it out

* feat(daily calculations):

arima.py:
* moved Repo inside class so the db connection is propery closed
when the Arima class gets deleted
* added an argument 'metric_ids' to get_data method signature in order
to decrease db transaction time

context.py:
* added a property to prediction to make generating the drop down list
more fluid

repository.py:
* alphabetized method order
* added new method to all runs as list rather than dataframe
* added 'metric_ids' as optional argument to get_measurements to reduce
the amount of unnecessary rows retrieved
* added a method 'get_predictions' to pull precomputed predictions from
the database to make ui more responsive
* changed method signature for put_measurements and added another method
so we can put measurements from either a csv file or a list

ui.py: new, a start to the ui

* feat(refactor): updates to daily.py

* feat(refactor): updates to daily.py

* feat(ui): distinguish between past and predicted values

* feat(ui): distinguish between past and predicted values

* feat(ui): distinguish between past and predicted values

* feat(ui): widget functionality

all widgets are working as intended. still need to comment and write a
few unit tests. i may try to write some selenium tests simply because
i never have and this is a small UI. would be a good place to try it out

* feat(daily calculations):

arima.py:
* moved Repo inside class so the db connection is propery closed
when the Arima class gets deleted
* added an argument 'metric_ids' to get_data method signature in order
to decrease db transaction time

context.py:
* added a property to prediction to make generating the drop down list
more fluid

repository.py:
* alphabetized method order
* added new method to all runs as list rather than dataframe
* added 'metric_ids' as optional argument to get_measurements to reduce
the amount of unnecessary rows retrieved
* added a method 'get_predictions' to pull precomputed predictions from
the database to make ui more responsive
* changed method signature for put_measurements and added another method
so we can put measurements from either a csv file or a list

ui.py: new, a start to the ui

* feat(refactor): updates to daily.py

* feat(refactor): updates to daily.py

* feat(ui): distinguish between past and predicted values

* feat(ui): distinguish between past and predicted values

* feat(ui): distinguish between past and predicted values

* feat(ui): widget functionality

all widgets are working as intended. still need to comment and write a
few unit tests. i may try to write some selenium tests simply because
i never have and this is a small UI. would be a good place to try it out

* .

* test(updated tests for current branch):